Initial commit - WIP
This is still very much work in progress. At the moment, this does not need any bug reports or feature ideas, there are a few unfinished things as of this commit, like mention counts (you always see 5 new messages and 5 new mentions) because I am still working on this every day, and since you cannot use it anyway until I post both the UnrealIRCd module and WordPress plugins, it's not recommended that you try this out yet, but you're free to nose around if you're curious, which is one of the reasons I'm posting this.
Not only that, I have my own ideas that I have barely implemented yet, and things are still not-ready for production.

TLDR: Unfinished, and unable to be used for now. This is just so you can see what's what, and so I can keep track of things more easily.